pass 2 on new strategy for engine module local tests
